refactor[#567] : 파일 컨벤션에 맞게 api 파일명 수정
